Check if method 1 everywhere-upper-dominates method 2

check_everywhere_dominance evaluates whether method 1 everywhere-upper-dominates method 2 at the tau-th quantile (Appendix Q.2).

Usage

check_everywhere_dominance(err1, err2, taulist = seq(0.5, 1, 0.01))

Arguments

err1

error object for method 1; see Details

err2

error object for method 2; see Details

taulist

a vector of quantiles at which the methods are compared

Value

a data.frame with two columns: the first column records the values of \(\tau\) and the second column records whether method 1 everywhere-upper-dominates method 2 at each quantile

Details

err should be in one of the following forms.

A square asymmetric matrix when the number of training domains \(r = 1\). In this case, err[i, j] records the transfer error using the i-th domain for training and j-th domain for testing.

A data.frame with \((r+2)\) columns. The last column records the transfer errors, the second last column records the indices of test domains, and the first r columns record the indices of training domains.

(Not recommended) a data.frame with two columns. The second column records the transfer errors and the first column records the indices of test domains.

For this version, it is recommended to include the transfer errors for all n-choose-(r+1) combinations of training and test domains. Otherwise, ete_interval can still output an interval, though the theoretical guarantee is unclear.

Examples

# Generate three error matrices such that
# errmat1 and errmat2 are similar 
# and errmat1 is smaller than errmat3
n <- 40
set.seed(1)
errmat1 <- matrix(runif(n^2), nrow = n)
errmat2 <- matrix(runif(n^2), nrow = n)
errmat3 <- errmat1 + 0.1
diag(errmat1) <- diag(errmat2) <- diag(errmat3) <- NA

check_everywhere_dominance(errmat1, errmat2)
#>     tau   res
#> 1  0.50 FALSE
#> 2  0.51 FALSE
#> 3  0.52 FALSE
#> 4  0.53 FALSE
#> 5  0.54 FALSE
#> 6  0.55 FALSE
#> 7  0.56 FALSE
#> 8  0.57 FALSE
#> 9  0.58 FALSE
#> 10 0.59 FALSE
#> 11 0.60 FALSE
#> 12 0.61 FALSE
#> 13 0.62 FALSE
#> 14 0.63 FALSE
#> 15 0.64 FALSE
#> 16 0.65 FALSE
#> 17 0.66 FALSE
#> 18 0.67 FALSE
#> 19 0.68 FALSE
#> 20 0.69 FALSE
#> 21 0.70 FALSE
#> 22 0.71 FALSE
#> 23 0.72 FALSE
#> 24 0.73 FALSE
#> 25 0.74 FALSE
#> 26 0.75 FALSE
#> 27 0.76 FALSE
#> 28 0.77 FALSE
#> 29 0.78 FALSE
#> 30 0.79 FALSE
#> 31 0.80 FALSE
#> 32 0.81 FALSE
#> 33 0.82 FALSE
#> 34 0.83 FALSE
#> 35 0.84 FALSE
#> 36 0.85 FALSE
#> 37 0.86 FALSE
#> 38 0.87 FALSE
#> 39 0.88 FALSE
#> 40 0.89 FALSE
#> 41 0.90 FALSE
#> 42 0.91 FALSE
#> 43 0.92 FALSE
#> 44 0.93 FALSE
#> 45 0.94 FALSE
#> 46 0.95 FALSE
#> 47 0.96 FALSE
#> 48 0.97 FALSE
#> 49 0.98 FALSE
#> 50 0.99 FALSE
#> 51 1.00 FALSE
check_everywhere_dominance(errmat1, errmat3)
#>     tau  res
#> 1  0.50 TRUE
#> 2  0.51 TRUE
#> 3  0.52 TRUE
#> 4  0.53 TRUE
#> 5  0.54 TRUE
#> 6  0.55 TRUE
#> 7  0.56 TRUE
#> 8  0.57 TRUE
#> 9  0.58 TRUE
#> 10 0.59 TRUE
#> 11 0.60 TRUE
#> 12 0.61 TRUE
#> 13 0.62 TRUE
#> 14 0.63 TRUE
#> 15 0.64 TRUE
#> 16 0.65 TRUE
#> 17 0.66 TRUE
#> 18 0.67 TRUE
#> 19 0.68 TRUE
#> 20 0.69 TRUE
#> 21 0.70 TRUE
#> 22 0.71 TRUE
#> 23 0.72 TRUE
#> 24 0.73 TRUE
#> 25 0.74 TRUE
#> 26 0.75 TRUE
#> 27 0.76 TRUE
#> 28 0.77 TRUE
#> 29 0.78 TRUE
#> 30 0.79 TRUE
#> 31 0.80 TRUE
#> 32 0.81 TRUE
#> 33 0.82 TRUE
#> 34 0.83 TRUE
#> 35 0.84 TRUE
#> 36 0.85 TRUE
#> 37 0.86 TRUE
#> 38 0.87 TRUE
#> 39 0.88 TRUE
#> 40 0.89 TRUE
#> 41 0.90 TRUE
#> 42 0.91 TRUE
#> 43 0.92 TRUE
#> 44 0.93 TRUE
#> 45 0.94 TRUE
#> 46 0.95 TRUE
#> 47 0.96 TRUE
#> 48 0.97 TRUE
#> 49 0.98 TRUE
#> 50 0.99 TRUE
#> 51 1.00 TRUE

# Transform error matrices into data.frames with r+2=3 columns
idx <- expand.grid(train = 1:n, test = 1:n)
errdf1 <- cbind(idx, data.frame(val = as.numeric(errmat1)))
errdf1 <- errdf1[!is.na(errdf1$val), ]
errdf2 <- cbind(idx, data.frame(val = as.numeric(errmat2)))
errdf2 <- errdf2[!is.na(errdf2$val), ]
errdf3 <- cbind(idx, data.frame(val = as.numeric(errmat3)))
errdf3 <- errdf3[!is.na(errdf3$val), ]

check_everywhere_dominance(errdf1, errdf2)
#>     tau   res
#> 1  0.50 FALSE
#> 2  0.51 FALSE
#> 3  0.52 FALSE
#> 4  0.53 FALSE
#> 5  0.54 FALSE
#> 6  0.55 FALSE
#> 7  0.56 FALSE
#> 8  0.57 FALSE
#> 9  0.58 FALSE
#> 10 0.59 FALSE
#> 11 0.60 FALSE
#> 12 0.61 FALSE
#> 13 0.62 FALSE
#> 14 0.63 FALSE
#> 15 0.64 FALSE
#> 16 0.65 FALSE
#> 17 0.66 FALSE
#> 18 0.67 FALSE
#> 19 0.68 FALSE
#> 20 0.69 FALSE
#> 21 0.70 FALSE
#> 22 0.71 FALSE
#> 23 0.72 FALSE
#> 24 0.73 FALSE
#> 25 0.74 FALSE
#> 26 0.75 FALSE
#> 27 0.76 FALSE
#> 28 0.77 FALSE
#> 29 0.78 FALSE
#> 30 0.79 FALSE
#> 31 0.80 FALSE
#> 32 0.81 FALSE
#> 33 0.82 FALSE
#> 34 0.83 FALSE
#> 35 0.84 FALSE
#> 36 0.85 FALSE
#> 37 0.86 FALSE
#> 38 0.87 FALSE
#> 39 0.88 FALSE
#> 40 0.89 FALSE
#> 41 0.90 FALSE
#> 42 0.91 FALSE
#> 43 0.92 FALSE
#> 44 0.93 FALSE
#> 45 0.94 FALSE
#> 46 0.95 FALSE
#> 47 0.96 FALSE
#> 48 0.97 FALSE
#> 49 0.98 FALSE
#> 50 0.99 FALSE
#> 51 1.00 FALSE
check_everywhere_dominance(errdf1, errdf3)
#>     tau  res
#> 1  0.50 TRUE
#> 2  0.51 TRUE
#> 3  0.52 TRUE
#> 4  0.53 TRUE
#> 5  0.54 TRUE
#> 6  0.55 TRUE
#> 7  0.56 TRUE
#> 8  0.57 TRUE
#> 9  0.58 TRUE
#> 10 0.59 TRUE
#> 11 0.60 TRUE
#> 12 0.61 TRUE
#> 13 0.62 TRUE
#> 14 0.63 TRUE
#> 15 0.64 TRUE
#> 16 0.65 TRUE
#> 17 0.66 TRUE
#> 18 0.67 TRUE
#> 19 0.68 TRUE
#> 20 0.69 TRUE
#> 21 0.70 TRUE
#> 22 0.71 TRUE
#> 23 0.72 TRUE
#> 24 0.73 TRUE
#> 25 0.74 TRUE
#> 26 0.75 TRUE
#> 27 0.76 TRUE
#> 28 0.77 TRUE
#> 29 0.78 TRUE
#> 30 0.79 TRUE
#> 31 0.80 TRUE
#> 32 0.81 TRUE
#> 33 0.82 TRUE
#> 34 0.83 TRUE
#> 35 0.84 TRUE
#> 36 0.85 TRUE
#> 37 0.86 TRUE
#> 38 0.87 TRUE
#> 39 0.88 TRUE
#> 40 0.89 TRUE
#> 41 0.90 TRUE
#> 42 0.91 TRUE
#> 43 0.92 TRUE
#> 44 0.93 TRUE
#> 45 0.94 TRUE
#> 46 0.95 TRUE
#> 47 0.96 TRUE
#> 48 0.97 TRUE
#> 49 0.98 TRUE
#> 50 0.99 TRUE
#> 51 1.00 TRUE

# Transform error matrices into data.frames with only two columns (not recommended)
errdf1_2 <- errdf1[, 2:3]
errdf2_2 <- errdf2[, 2:3]
errdf3_2 <- errdf3[, 2:3]

check_everywhere_dominance(errdf1_2, errdf2_2)
#>     tau   res
#> 1  0.50 FALSE
#> 2  0.51 FALSE
#> 3  0.52 FALSE
#> 4  0.53 FALSE
#> 5  0.54 FALSE
#> 6  0.55 FALSE
#> 7  0.56 FALSE
#> 8  0.57 FALSE
#> 9  0.58 FALSE
#> 10 0.59 FALSE
#> 11 0.60 FALSE
#> 12 0.61 FALSE
#> 13 0.62 FALSE
#> 14 0.63 FALSE
#> 15 0.64 FALSE
#> 16 0.65 FALSE
#> 17 0.66 FALSE
#> 18 0.67 FALSE
#> 19 0.68 FALSE
#> 20 0.69 FALSE
#> 21 0.70 FALSE
#> 22 0.71 FALSE
#> 23 0.72 FALSE
#> 24 0.73 FALSE
#> 25 0.74 FALSE
#> 26 0.75 FALSE
#> 27 0.76 FALSE
#> 28 0.77 FALSE
#> 29 0.78 FALSE
#> 30 0.79 FALSE
#> 31 0.80 FALSE
#> 32 0.81 FALSE
#> 33 0.82 FALSE
#> 34 0.83 FALSE
#> 35 0.84 FALSE
#> 36 0.85 FALSE
#> 37 0.86 FALSE
#> 38 0.87 FALSE
#> 39 0.88 FALSE
#> 40 0.89 FALSE
#> 41 0.90 FALSE
#> 42 0.91 FALSE
#> 43 0.92 FALSE
#> 44 0.93 FALSE
#> 45 0.94 FALSE
#> 46 0.95 FALSE
#> 47 0.96 FALSE
#> 48 0.97 FALSE
#> 49 0.98 FALSE
#> 50 0.99 FALSE
#> 51 1.00 FALSE
check_everywhere_dominance(errdf1_2, errdf3_2)
#>     tau  res
#> 1  0.50 TRUE
#> 2  0.51 TRUE
#> 3  0.52 TRUE
#> 4  0.53 TRUE
#> 5  0.54 TRUE
#> 6  0.55 TRUE
#> 7  0.56 TRUE
#> 8  0.57 TRUE
#> 9  0.58 TRUE
#> 10 0.59 TRUE
#> 11 0.60 TRUE
#> 12 0.61 TRUE
#> 13 0.62 TRUE
#> 14 0.63 TRUE
#> 15 0.64 TRUE
#> 16 0.65 TRUE
#> 17 0.66 TRUE
#> 18 0.67 TRUE
#> 19 0.68 TRUE
#> 20 0.69 TRUE
#> 21 0.70 TRUE
#> 22 0.71 TRUE
#> 23 0.72 TRUE
#> 24 0.73 TRUE
#> 25 0.74 TRUE
#> 26 0.75 TRUE
#> 27 0.76 TRUE
#> 28 0.77 TRUE
#> 29 0.78 TRUE
#> 30 0.79 TRUE
#> 31 0.80 TRUE
#> 32 0.81 TRUE
#> 33 0.82 TRUE
#> 34 0.83 TRUE
#> 35 0.84 TRUE
#> 36 0.85 TRUE
#> 37 0.86 TRUE
#> 38 0.87 TRUE
#> 39 0.88 TRUE
#> 40 0.89 TRUE
#> 41 0.90 TRUE
#> 42 0.91 TRUE
#> 43 0.92 TRUE
#> 44 0.93 TRUE
#> 45 0.94 TRUE
#> 46 0.95 TRUE
#> 47 0.96 TRUE
#> 48 0.97 TRUE
#> 49 0.98 TRUE
#> 50 0.99 TRUE
#> 51 1.00 TRUE

# Generate an error data.frame with r>1
n <- 20
set.seed(1)
idx <- expand.grid(train1 = 1:n, train2 = 1:n, test = 1:n)
idx <- idx[idx$train1 != idx$train2 & idx$train1 != idx$test & idx$train2 != idx$test, ]
err1 <- cbind(idx, data.frame(val = runif(nrow(idx))))
err2 <- cbind(idx, data.frame(val = runif(nrow(idx))))
err3 <- cbind(idx, data.frame(val = err1$val + 0.1))

check_everywhere_dominance(err1, err2)
#>     tau   res
#> 1  0.50 FALSE
#> 2  0.51 FALSE
#> 3  0.52 FALSE
#> 4  0.53 FALSE
#> 5  0.54 FALSE
#> 6  0.55 FALSE
#> 7  0.56 FALSE
#> 8  0.57 FALSE
#> 9  0.58 FALSE
#> 10 0.59 FALSE
#> 11 0.60 FALSE
#> 12 0.61 FALSE
#> 13 0.62 FALSE
#> 14 0.63 FALSE
#> 15 0.64 FALSE
#> 16 0.65 FALSE
#> 17 0.66 FALSE
#> 18 0.67 FALSE
#> 19 0.68 FALSE
#> 20 0.69 FALSE
#> 21 0.70 FALSE
#> 22 0.71 FALSE
#> 23 0.72 FALSE
#> 24 0.73 FALSE
#> 25 0.74 FALSE
#> 26 0.75 FALSE
#> 27 0.76 FALSE
#> 28 0.77 FALSE
#> 29 0.78 FALSE
#> 30 0.79 FALSE
#> 31 0.80 FALSE
#> 32 0.81 FALSE
#> 33 0.82 FALSE
#> 34 0.83 FALSE
#> 35 0.84 FALSE
#> 36 0.85 FALSE
#> 37 0.86 FALSE
#> 38 0.87 FALSE
#> 39 0.88 FALSE
#> 40 0.89 FALSE
#> 41 0.90 FALSE
#> 42 0.91 FALSE
#> 43 0.92 FALSE
#> 44 0.93 FALSE
#> 45 0.94 FALSE
#> 46 0.95 FALSE
#> 47 0.96 FALSE
#> 48 0.97 FALSE
#> 49 0.98 FALSE
#> 50 0.99 FALSE
#> 51 1.00 FALSE
check_everywhere_dominance(err1, err3)
#>     tau  res
#> 1  0.50 TRUE
#> 2  0.51 TRUE
#> 3  0.52 TRUE
#> 4  0.53 TRUE
#> 5  0.54 TRUE
#> 6  0.55 TRUE
#> 7  0.56 TRUE
#> 8  0.57 TRUE
#> 9  0.58 TRUE
#> 10 0.59 TRUE
#> 11 0.60 TRUE
#> 12 0.61 TRUE
#> 13 0.62 TRUE
#> 14 0.63 TRUE
#> 15 0.64 TRUE
#> 16 0.65 TRUE
#> 17 0.66 TRUE
#> 18 0.67 TRUE
#> 19 0.68 TRUE
#> 20 0.69 TRUE
#> 21 0.70 TRUE
#> 22 0.71 TRUE
#> 23 0.72 TRUE
#> 24 0.73 TRUE
#> 25 0.74 TRUE
#> 26 0.75 TRUE
#> 27 0.76 TRUE
#> 28 0.77 TRUE
#> 29 0.78 TRUE
#> 30 0.79 TRUE
#> 31 0.80 TRUE
#> 32 0.81 TRUE
#> 33 0.82 TRUE
#> 34 0.83 TRUE
#> 35 0.84 TRUE
#> 36 0.85 TRUE
#> 37 0.86 TRUE
#> 38 0.87 TRUE
#> 39 0.88 TRUE
#> 40 0.89 TRUE
#> 41 0.90 TRUE
#> 42 0.91 TRUE
#> 43 0.92 TRUE
#> 44 0.93 TRUE
#> 45 0.94 TRUE
#> 46 0.95 TRUE
#> 47 0.96 TRUE
#> 48 0.97 TRUE
#> 49 0.98 TRUE
#> 50 0.99 TRUE
#> 51 1.00 TRUE