Forecast intervals for transfer errors

forecast_interval generates one- or two-sided forecast intervals for transfer errors (Section 3) with a given coverage level with or without finite-sample correction.

Usage

forecast_interval(
  err,
  coverage,
  side = c("two", "left", "right"),
  correction = FALSE
)

Arguments

err

error object; see Details

coverage

target coverage level

side

"two" for two-sided, "left" for intervals of form \([a, \infty)\), "right" for intervals of form \((-\infty, a]\)

correction

TRUE if correction is performed to guarantee the coverage is below alpha in finite samples; FALSE if not correction is performed; see Details

Value

an interval as a vector of length 2

Details

err should be in one of the following forms.

A square asymmetric matrix when the number of training domains \(r = 1\). In this case, err[i, j] records the transfer error using the i-th domain for training and j-th domain for testing.

A data.frame with \((r+2)\) columns. The last column records the transfer errors, the second last column records the indices of test domains, and the first r columns record the indices of training domains.

(Not recommended) a data.frame with two columns. The second column records the transfer errors and the first column records the indices of test domains.

For this version, it is recommended to include the transfer errors for all n-choose-(r+1) combinations of training and test domains. Otherwise, forecast_interval can still output an interval, though the theoretical guarantee is unclear. When correction = FALSE, the output is

\([\underline{e}_{(1+\tau)/2}^{\mathbf{M}}, \overline{e}_{(1+\tau)/2}^{\mathbf{M}}]\) when side = "two";

\([\underline{e}_{\tau}^{\mathbf{M}}, \infty)\) when side = "left";

\((-\infty, \overline{e}_{\tau}^{\mathbf{M}}]\) when side = "right",

where \(\tau\) is given by coverage. The theoretical lower bound for coverage is smaller than \(\tau\) (Proposition 1) with the gap \(O(r/n)\).

When correction = TRUE, it first multiplies coverage by \((n + 1) / (n - r)\) and then outputs the above interval. Proposition 1 implies that the coverage is lower bounded by \(\tau\) in finite samples.

Examples

# Generate an error matrix
n <- 100
set.seed(1)
errmat <- matrix(runif(n^2), nrow = n)
diag(errmat) <- NA

# Two-sided forecast intervals without and with finite-sample correction
forecast_interval(errmat, 0.9, "two", correction = FALSE)
#> [1] 0.04861056 0.95264364
forecast_interval(errmat, 0.9, "two", correction = TRUE)
#> [1] 0.03910006 0.96040231

# One-sided forecast intervals without and with finite-sample correction
forecast_interval(errmat, 0.9, "left", correction = FALSE)
#> [1] 0.09737435        Inf
forecast_interval(errmat, 0.9, "left", correction = TRUE)
#> [1] 0.07821896        Inf
forecast_interval(errmat, 0.9, "right", correction = FALSE)
#> [1]      -Inf 0.9039679
forecast_interval(errmat, 0.9, "right", correction = TRUE)
#> [1]      -Inf 0.9219315

# Generate an error data.frame with r+2=3 columns
idx <- expand.grid(train = 1:n, test = 1:n)
errdf <- cbind(idx, data.frame(val = as.numeric(errmat)))
errdf <- errdf[!is.na(errdf$val), ]
forecast_interval(errdf, 0.9, "two", correction = FALSE)
#> [1] 0.04861056 0.95264364
forecast_interval(errmat, 0.9, "two", correction = FALSE)
#> [1] 0.04861056 0.95264364

# Generate an error data.frame with only two columns (not recommended)
errdf2 <- errdf[, 2:3]
forecast_interval(errdf2, 0.9, "two", correction = FALSE)
#> [1] 0.04861056 0.95264364

# Generate an error data.frame with r>1
n <- 40
set.seed(1)
idx <- expand.grid(train1 = 1:n, train2 = 1:n, test = 1:n)
idx <- idx[idx$train1 != idx$train2 & idx$train1 != idx$test & idx$train2 != idx$test, ]
err <- cbind(idx, data.frame(val = runif(nrow(idx))))

# Two-sided forecast intervals without and with finite-sample correction
forecast_interval(err, 0.9, "two", correction = FALSE)
#> [1] 0.04999229 0.95099817
forecast_interval(err, 0.9, "two", correction = TRUE)
#> [1] 0.01426661 0.98641024

# One-sided forecast intervals without and with finite-sample correction
forecast_interval(err, 0.9, "left", correction = FALSE)
#> [1] 0.09928197        Inf
forecast_interval(err, 0.9, "left", correction = TRUE)
#> [1] 0.02866646        Inf
forecast_interval(err, 0.9, "right", correction = FALSE)
#> [1]      -Inf 0.9009274
forecast_interval(err, 0.9, "right", correction = TRUE)
#> [1]     -Inf 0.972384